refactor: move share link button comp to Dashboard folder
